requirements.txt
numpy
pandas
scikit-learn
nltk
matplotlib
seaborn
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
deep-translator
deep-translator
torch
torchvision
torchaudio
transformers
scikit-learn
pillow
opencv-python
